+++ HISTORY.rst
@@ -3,7 +3,14 @@
 History
 -------
 
-0.1.0 (2015-01-11)
+0.2.0 (2015-07-28)
+---------------------
+
+* Cleaned up with parts rewritten.
+* Takes files as command line arguments.
+* Handles mutiple images.
+
+0.1.0 (2015-06-06)
 ---------------------
 
 * First release on PyPI.